Исправлен график, убраны все циклы (вродебы), если не правильные значения задали в массивк то об этом скажет

+++ work_with_file.py
@@ -3,7 +3,7 @@ import numpy as np
 def values_from_file(path: str = None) -> np.ndarray:
     try:
         with open(path) as f:
-            return np.loadtxt(f, dtype=np.int8)
+            return np.loadtxt(f)
             # return values if len(values) != 0 else None
     except FileNotFoundError:
         return None
